# [openrr-planner] Use AABB-based pruning in self-collision checking