--- a/.github/actions/spelling/expect.txt
+++ b/.github/actions/spelling/expect.txt
@@ -1,3 +1,4 @@
+aabb
 abi
 actionlib
 adduser
@@ -13,6 +14,8 @@ attr
 backend
 bringup
 buf
+BVH
+BVT
 cdylib
 cfg
 checkbox