This PR implements a sweep and prune algorithm with AABB (Axis Aligned Bounding Box), in order to accelerate self-collision checking. This implementation is based on bounding volume trees provided by ncollide3d (see https://www.ncollide.org/bounding_volumes/#spacial-partitioning ).

We can evaluate the improvement in computational time with `cargo bench`.
The results in my environment are as follows:

```
[Before]
bench_check_self_collisions                        
                        time:   [4.6853 ms 4.7762 ms 4.8687 ms]

[After (this PR is applied)]
bench_check_self_collisions                        
                        time:   [177.70 us 179.92 us 182.28 us]
                        change: [-96.318% -96.233% -96.146%] (p = 0.00 < 0.05)
                        Performance has improved.
```